server/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
httpx>=0.27.0
openenv-core==0.2.3
